Make LDAP search filter configurable

+++ app/logger.js
@@ -1,6 +1,6 @@
 const bunyan = require('bunyan');
 const bformat = require('bunyan-format');
-const formatOut = bformat({ outputMode: 'short', color: true });
+const formatOut = bformat({ outputMode: 'long', color: true });
 
 const logger = bunyan.createLogger({ // also used for logging in ldapauth-fork and ldapjs
 	name: 'ldap-jwt',
